Вынесена логика приложения из ConsoleUI
