+++ TagsCloudContainer/TagsCloudContainer.csproj
@@ -41,6 +41,7 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Core\ApplicationCore.cs" />
     <Compile Include="Core\ColoringAlgorithms\IColoringAlgorithm.cs" />
     <Compile Include="Core\ColoringAlgorithms\RandomColoring.cs" />
     <Compile Include="Core\Extensions\IEnumerableExtensions.cs" />
